Day 16 - Visual studio compiler switches

+++ handmade/code/handmade.cpp
@@ -34,8 +34,8 @@ void RenderWeirdGradient(game_offscreen_buffer* Buffer, int BlueOffset, int Gree
         uint32* Pixel = (uint32*)Row;
         for (int X = 0; X < Buffer->Width; ++X)
         {
-            uint8 Blue = (X + BlueOffset);
-            uint8 Green = (Y + GreenOffset);
+            uint8 Blue = (uint8)(X + BlueOffset);
+            uint8 Green = (uint8)(Y + GreenOffset);
 
             *Pixel++ = ((Green << 8) | Blue);
         }
@@ -72,8 +72,8 @@ void GameUpdateAndRender(game_memory *Memory, game_input* Input,
     if (Input0->IsAnalog)
     {
         // Use analog movement tuning
-        GameState->BlueOffset += (int)4.0f * (Input0->EndX);
-        GameState->ToneHz = 256 + (int)(128.0f *(Input0->EndY));
+        GameState->BlueOffset += (int)(4.0f * Input0->EndX);
+        GameState->ToneHz = 256 + (int)(128.0f *Input0->EndY);
     }
     else
     {
